+++ framework-4.2_phpfmt/src/Illuminate/Queue/Worker.php
@@ -112,11 +112,7 @@ class Worker {
 		}
 		catch (\Exception $e)
 		{
-			if ($this->exceptions)
-			{
-				$this->exceptions->handleException($e);
-			}
-
+			if ($this->exceptions) $this->exceptions->handleException($e);
 		}
 	}
 
@@ -175,18 +171,11 @@ class Worker {
 	 */
 	protected function getNextJob($connection, $queue)
 	{
-		if (is_null($queue))
-		{
-			return $connection->pop();
-		}
+		if (is_null($queue)) return $connection->pop();
 
 		foreach (explode(',', $queue) as $queue)
 		{
-			if ( ! is_null($job = $connection->pop($queue)))
-			{
-				return $job;
-			}
-
+			if ( ! is_null($job = $connection->pop($queue))) return $job;
 		}
 	}
 
@@ -215,10 +204,7 @@ class Worker {
 			// the delete method on the job. Otherwise we will just keep moving.
 			$job->fire();
 
-			if ($job->autoDelete())
-			{
-				$job->delete();
-			}
+			if ($job->autoDelete()) $job->delete();
 
 			return ['job' => $job, 'failed' => false];
 		}
@@ -227,10 +213,7 @@ class Worker {
 			// If we catch an exception, we will attempt to release the job back onto
 			// the queue so it is not lost. This will let is be retried at a later
 			// time by another listener (or the same one). We will do that here.
-			if ( ! $job->isDeleted())
-			{
-				$job->release($delay);
-			}
+			if ( ! $job->isDeleted()) $job->release($delay);
 
 			throw $e;
 		}
